move polyfill to separate file
